+++ src/inscription-parser.service.envelope.spec.ts
@@ -1,5 +1,6 @@
 import { InscriptionParserService } from './inscription-parser.service';
-import { getNextInscriptionMark, hexStringToUint8Array } from './inscription-parser.service.helper';
+import { getNextInscriptionMark } from './inscription-parser.service.helper';
+import { hexStringToUint8Array } from './lib/conversions';
 import { readInscriptionAsBase64, readTransaction } from './test.helper';
 
 describe('Inscription parser', () => {
