+++ src/inscription-parser.service.helper.ts
@@ -50,54 +50,6 @@ export function getKnownFieldValue(fields: { tag: Uint8Array; value: Uint8Array 
   return knownField.value;
 }
 
-/**
- * Converts a hex string to a Uint8Array.
- *
- * @param hexStr - The hex string to be converted.
- * @returns The resulting Uint8Array.
- */
-export function hexStringToUint8Array(hex: string): Uint8Array {
-  if (hex.length === 0) {
-    throw new Error('Input string is empty. Hex string expected.');
-  }
-  return new Uint8Array(hex.match(/.{1,2}/g)!.map(byte => parseInt(byte, 16)));
-}
-
-/**
- * Converts a Uint8Array containing UTF-8 encoded data to a normal a UTF-16 encoded string.
- *
- * @param bytes - The Uint8Array containing UTF-8 encoded data.
- * @returns The corresponding UTF-16 encoded JavaScript string.
- */
-export function utf8BytesToUtf16String(bytes: Uint8Array | undefined): string | undefined {
-
-  if (bytes === undefined) {
-    return undefined;
-  }
-
-  const decoder = new TextDecoder('utf-8');
-  return decoder.decode(bytes);
-}
-
-/**
- * Convert a Uint8Array to a string by treating each byte as a character code.
- * It avoids interpreting bytes as UTF-8 encoded sequences.
- * --> Again: it ignores UTF-8 encoding, which is necessary for binary content!
- *
- * Note: This method is different from using `String.fromCharCode(...combinedData)` which can
- * cause a "Maximum call stack size exceeded" error for large arrays due to the limitation of
- * the spread operator in JavaScript. (previously the parser broke here, because of large content)
- *
- * @param bytes - The byte array to convert.
- * @returns The resulting string where each byte value is treated as a direct character code.
- */
-export function uint8ArrayToSingleByteChars(bytes: Uint8Array): string {
-  let resultStr = '';
-  for (let i = 0; i < bytes.length; i++) {
-    resultStr += String.fromCharCode(bytes[i]);
-  }
-  return resultStr;
-}
 
 /**
  * Reads a specified number of bytes from a Uint8Array starting from a given pointer.
